Vehiculos

+++ DGT/DGT.Web/Controllers/InfraccionesController.cs
@@ -19,29 +19,10 @@ namespace DGT.Web.Controllers
         public object Get(int page = 1, int pageSize = 5)
         {
             IQueryable<Infraccion> query;
-            page = page - 1;
+
             query = TheRepository.GetAllInfracciones().OrderBy(c => c.Puntos);
-            var totalCount = query.Count();
-            var totalPages = (int)Math.Ceiling((double)totalCount / pageSize);
 
-            var urlHelper = new UrlHelper(Request);
-            var prevLink = page > 0 ? urlHelper.Link("Infracciones", new { page = page - 1, pageSize = pageSize }) : "";
-            var nextLink = page < totalPages - 1 ? urlHelper.Link("Infracciones", new { page = page + 1, pageSize = pageSize }) : "";
-
-            var results = query
-                          .Skip(pageSize * page)
-                          .Take(pageSize)
-                          .ToList()
-                          .Select(s => TheModelFactory.Create(s));
-
-            return new
-            {
-                TotalCount = totalCount,
-                TotalPages = totalPages,
-                PrevPageLink = prevLink,
-                NextPageLink = nextLink,
-                Results = results
-            };
+            return CreateResponse(query, page, pageSize, "Infracciones");
 
         }
 
@@ -49,7 +30,6 @@ namespace DGT.Web.Controllers
         {
             try
             {
-
                 if (TheRepository.Insert(infraccion) && TheRepository.SaveAll())
                 {
                     return Request.CreateResponse(HttpStatusCode.Created, TheModelFactory.Create(infraccion));
@@ -68,28 +48,28 @@ namespace DGT.Web.Controllers
 
         [HttpPatch]
         [HttpPut]
-        public HttpResponseMessage Put(string dni, [FromBody] Infraccion infraccion)
+        public HttpResponseMessage Put(int id, [FromBody] Infraccion infraccion)
         {
             try
             {
 
-                var infraccionOriginal = TheRepository.GetAllInfraccionesByConductor(dni);
+                var infraccionOriginal = TheRepository.GetInfraccionById(id).FirstOrDefault();
 
-                if (infraccionOriginal == null || infraccionOriginal.Count() < 1)
+                if (infraccionOriginal == null)
                 {
-                    return Request.CreateResponse(HttpStatusCode.NotModified, "No se encuentra");
+                    return Request.CreateResponse(HttpStatusCode.NotModified, "No se encuentra la infracción");
                 }
-                //infraccion.Id = infraccionOriginal.Id;
 
-                //if (TheRepository.Update(infraccionOriginal, infraccion) && TheRepository.SaveAll())
-                //{
-                //    return Request.CreateResponse(HttpStatusCode.OK, TheModelFactory.Create(infraccion));
-                //}
-                //else
-                //{
-                //    return Request.CreateResponse(HttpStatusCode.NotModified);
-                //}
-                return Request.CreateResponse(HttpStatusCode.NotModified);
+                infraccion.Id = infraccionOriginal.Id;
+
+                if (TheRepository.Update(infraccionOriginal, infraccion) && TheRepository.SaveAll())
+                {
+                    return Request.CreateResponse(HttpStatusCode.OK, TheModelFactory.Create(infraccion));
+                }
+                else
+                {
+                    return Request.CreateResponse(HttpStatusCode.NotModified);
+                }
             }
             catch (Exception ex)
             {
@@ -101,17 +81,17 @@ namespace DGT.Web.Controllers
         {
             try
             {
-                var infraccion = TheRepository.GetAllInfraccionesByConductor(id.ToString());
+                var infraccion = TheRepository.GetInfraccionById(id);
 
                 if (infraccion == null)
                 {
                     return Request.CreateResponse(HttpStatusCode.NotFound);
                 }
-
-                //if (infraccion.Count > 0)
-                //{
-                //    return Request.CreateResponse(HttpStatusCode.BadRequest, "No se puede eliminar infracción.");
-                //}
+                var conductores = TheRepository.GetAllConductoresByInfraccion(id);
+                if (conductores.Count() > 0)
+                {
+                    return Request.CreateResponse(HttpStatusCode.BadRequest, "No se puede eliminar infracción porque hay conductores con esta infracción.");
+                }
 
                 if (TheRepository.DeleteInfraccion(id) && TheRepository.SaveAll())
                 {
